u64 to usize + Wrote queueing network connectors

+++ src/queues/poisson_generator.rs
@@ -5,14 +5,14 @@ use self::rand::distributions::Sample;
 use queues::request::Request;
 use queues::Queue;
 
-pub struct PoissonGenerator<T> where T: Sample<u64> {
+pub struct PoissonGenerator<T> where T: Sample<usize> {
     next_exit: f64,
     next_request: Option<Request>,
     pop_distribution: T,
     iat_distribution: Exp
 }
 
-impl<T> Queue for PoissonGenerator<T> where T: Sample<u64> {
+impl<T> Queue for PoissonGenerator<T> where T: Sample<usize> {
     fn arrival (&mut self, _req: Request) {
         panic!("You should not arrive at a generator");
     }
@@ -36,7 +36,7 @@ impl<T> Queue for PoissonGenerator<T> where T: Sample<u64> {
     }
 }
 
-impl<T> PoissonGenerator<T> where T: Sample<u64> {
+impl<T> PoissonGenerator<T> where T: Sample<usize> {
     pub fn new (rate: f64, distribution: T) -> Self{
         let mut ret = PoissonGenerator {
             next_exit: 0.,
