+++ src/queues/zipfgen.rs
@@ -40,7 +40,7 @@ impl<T> ZipfGenerator<T> where T: Sample<f64> {
 
     fn insert_exit(&mut self, content: usize, time: f64) {
         let ntime = time + self.iat_distribution[content-1].sample(&mut rand::thread_rng());
-        self.next_arrivals.push(ntime, Request::new(content as u64));
+        self.next_arrivals.push(ntime, Request::new(content));
     }
 
     pub fn get_alpha (&self) -> f64 {
